+++ src/ScottPlot5/ScottPlot5/Plot.cs
@@ -9,7 +9,6 @@ public class Plot : IDisposable
 {
     public List<IXAxis> XAxes { get; } = new(); // TODO: axes should be inside the Panels list
     public List<IYAxis> YAxes { get; } = new(); // TODO: axes should be inside the Panels list
-    public List<IAxis> Axes => Enumerable.Concat<IAxis>(XAxes, YAxes).ToList();
 
     public IXAxis TopAxis => XAxes.First(x => x.Edge == Edge.Top);
     public IXAxis BottomAxis => XAxes.First(x => x.Edge == Edge.Bottom);